requirements.txt
django==1.11.20
django-crispy-forms==1.7.2
django-import-export>=1.2.0
django-reversion==3.0.2
django-formtools==2.1
future==0.17.1
httplib2==0.12.0
six==1.11.0
djangorestframework==3.7.7
wechatpy==1.8.1
paho-mqtt==1.4.0
pillow==5.4.1
qrcode==6.1
aliyun-python-sdk-core==2.12
pymysql==0.9.3
channels==2.1.7
channels-redis==2.3.3
pywin32==224
websockets==7.0
redis==3.2.1
pycryptodome==3.8.1



